moved data-related helper functions to data.py
