+++ joeynmt/helpers.py
@@ -6,7 +6,6 @@ import copy
 import glob
 import os
 import os.path
-import sys
 from collections import Counter
 from logging import Logger
 from typing import Callable
@@ -197,113 +196,6 @@ def log_data_info(train_data: Dataset, valid_data: Dataset, test_data: Dataset,
 
     logging_function("Number of Src words (types): %d", len(src_vocab))
     logging_function("Number of Trg words (types): %d", len(trg_vocab))
-
-
-def load_data(cfg):
-    """
-    Load train, dev and test data as specified in ccnfiguration.
-
-    :param cfg:
-    :return:
-    """
-    # load data from files
-    data_cfg = cfg["data"]
-    src_lang = data_cfg["src"]
-    trg_lang = data_cfg["trg"]
-    train_path = data_cfg["train"]
-    dev_path = data_cfg["dev"]
-    test_path = data_cfg.get("test", None)
-    level = data_cfg["level"]
-    lowercase = data_cfg["lowercase"]
-    max_sent_length = data_cfg["max_sent_length"]
-
-    #pylint: disable=unnecessary-lambda
-    if level == "char":
-        tok_fun = lambda s: list(s)
-    else:  # bpe or word, pre-tokenized
-        tok_fun = lambda s: s.split()
-
-    src_field = data.Field(init_token=None, eos_token=EOS_TOKEN,
-                           pad_token=PAD_TOKEN, tokenize=tok_fun,
-                           batch_first=True, lower=lowercase,
-                           unk_token=UNK_TOKEN,
-                           include_lengths=True)
-
-    trg_field = data.Field(init_token=BOS_TOKEN, eos_token=EOS_TOKEN,
-                           pad_token=PAD_TOKEN, tokenize=tok_fun,
-                           unk_token=UNK_TOKEN,
-                           batch_first=True, lower=lowercase,
-                           include_lengths=True)
-
-    train_data = TranslationDataset(path=train_path,
-                                    exts=("." + src_lang, "." + trg_lang),
-                                    fields=(src_field, trg_field),
-                                    filter_pred=
-                                    lambda x: len(vars(x)['src'])
-                                    <= max_sent_length
-                                    and len(vars(x)['trg'])
-                                    <= max_sent_length)
-
-    max_size = data_cfg.get("voc_limit", sys.maxsize)
-    min_freq = data_cfg.get("voc_min_freq", 1)
-    src_vocab_file = data_cfg.get("src_vocab", None)
-    trg_vocab_file = data_cfg.get("trg_vocab", None)
-
-    src_vocab = build_vocab(field="src", min_freq=min_freq, max_size=max_size,
-                            dataset=train_data, vocab_file=src_vocab_file)
-    trg_vocab = build_vocab(field="trg", min_freq=min_freq, max_size=max_size,
-                            dataset=train_data, vocab_file=trg_vocab_file)
-    dev_data = TranslationDataset(path=dev_path,
-                                  exts=("." + src_lang, "." + trg_lang),
-                                  fields=(src_field, trg_field))
-    test_data = None
-    if test_path is not None:
-        # check if target exists
-        if os.path.isfile(test_path + "." + trg_lang):
-            test_data = TranslationDataset(
-                path=test_path, exts=("." + src_lang, "." + trg_lang),
-                fields=(src_field, trg_field))
-        else:
-            # no target is given -> create dataset from src only
-
-            test_data = MonoDataset(path=test_path, ext="." + src_lang,
-                                    field=(src_field))
-    src_field.vocab = src_vocab
-    trg_field.vocab = trg_vocab
-    return train_data, dev_data, test_data, src_vocab, trg_vocab
-
-
-class MonoDataset(Dataset):
-    """Defines a dataset for machine translation without targets."""
-
-
-    @staticmethod
-    def sort_key(ex):
-        return len(ex.src)
-
-    def __init__(self, path, ext, field, **kwargs):
-        """Create a MonoDataset given path and field.
-
-        Arguments:
-            path: Prefix of path to the data file
-            ext: Containing the extension to path for this language.
-            field: Containing the fields that will be used for data
-            Remaining keyword arguments: Passed to the constructor of
-                data.Dataset.
-        """
-        fields = [('src', field)]
-
-        src_path = os.path.expanduser(path + ext)
-
-        examples = []
-        with open(src_path) as src_file:
-            for src_line in src_file:
-                src_line = src_line.strip()
-                if src_line != '':
-                    examples.append(data.Example.fromlist(
-                        [src_line], fields))
-
-        super(MonoDataset, self).__init__(examples, fields, **kwargs)
 
 
 def load_config(path="configs/default.yaml"):
